Move SourceTypes to Subject
Closes #292
