+++ src/app/store/index.ts
@@ -9,16 +9,13 @@ import { storeFreeze } from 'ngrx-store-freeze'
 
 import { ENV } from '../../environments/environment'
 import { CustomRouterState } from '../shared/utils/custom-router-state-serializer'
-import * as fromSourceTypes from './source-types/source-types.reducer'
 
 export interface State {
   router: fromRouter.RouterReducerState<CustomRouterState>
-  sourceTypes: fromSourceTypes.State
 }
 
 export const reducers: ActionReducerMap<State> = {
-  router: fromRouter.routerReducer,
-  sourceTypes: fromSourceTypes.reducer
+  router: fromRouter.routerReducer
 }
 
 export const metaReducers: MetaReducer<{}>[] = !ENV.PROD ? [storeFreeze] : []
@@ -46,14 +43,3 @@ export const getRouterParamsSubjectId = createSelector(
   getRouterState,
   (router): string => router && router.state.params['subjectId']
 )
-
-export const getSourceTypesState = createFeatureSelector<fromSourceTypes.State>(
-  'sourceTypes'
-)
-
-export const {
-  selectIds: getSourceTypesIds,
-  selectEntities: getSourceTypesEntities,
-  selectAll: getSourceTypesAll,
-  selectTotal: getSourceTypesTotal
-} = fromSourceTypes.adapter.getSelectors(getSourceTypesState)
