+++ src/app/subject/services/source-types.service.ts
@@ -1,7 +1,6 @@
 import { HttpClient } from '@angular/common/http'
 import { Injectable } from '@angular/core'
 import { Observable } from 'rxjs'
-import { tap } from 'rxjs/operators'
 
 import { ENV } from '../../../environments/environment'
 import { SourceType } from '../../shared/models/source-type.model'
